+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## v5.4.0-xx - xxxx-xx-xx
 
+### Bug Fixes:
 
+* Fixed transparent theme colors
+* Updated colors properly when theme changes
+* Added some more carrier cargo sanity checks
 
 
 ## v5.4.0-a2 - 2025-12-08